+++ sijinn-8/src/main/java/it/sijinn/perceptron/functions/applied/SimpleSigmoidFermi.java
@@ -23,7 +23,7 @@ public class SimpleSigmoidFermi implements IFunctionApplied {
 	@Override
 	public float execution(float[] param) {
 		if(param.length>0)
-			return new Float(1/(1+Math.pow(Math.E, -1*param[0]))).floatValue();
+			return new Float(1/(1+Math.pow(Math.E, -1*param[0])));
 		return 0;
 	}
